Do not aq_base() the underlaying listing object when delegating attribute access.
The check for attribute existence should be done over on aq_base()'d
object but when the existence is determined the actual attribute
access should be done using the acquisition wrapped object to not
break calls to methods that rely on acquisition.

For example

    name = 'absolute_url'
    if hasattr(aq_base(obj), name):
        return getattr(ab_base(obj), name)

fails since obj.absolute_url() does not work without acquisition.

    name = 'absolute_url'
    if hasattr(aq_base(obj), name):
        return getattr(obj, name)

allows the call to succeed.